web: shift all prisma stuff to nextauth
